requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg==3.2.3
psycopg-binary==3.2.3
PyJWT==2.10.1
sqlparse==0.5.2
tzdata==2024.2
